Simplify menu creation in AnimalTrapBlockEntity
The AnimalTrapMenu instance is now being created with 'this' as the argument instead of writing the block position into a new FriendlyByteBuf. This simplifies the code and makes it easier to understand. Also, the usage of the interaction result has been altered to CONSUME in the use() method in the AnimalTrapBlock class to absorb the action. The opening of the menu container has been streamlined as well.